cmd/snap-seccomp: add 'version-info' command
Add a sub-command to obtain a version information of snap-seccomp and the
underlying libseccomp library.

The version information combines the version reported by libseccomp with a hash
of all syscall names supported by the library. The reason is to be able to catch
the scenario in which libseccomp reports the same verbion, but due to distro
applied patches, the list of supported syscalls can change.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>