interfaces/seccomp: combine snap-seccomp version information with the source profile
As a way of detecting when we need to regenerate the BPF profile, try to include
snap-seccomp version information in the source version of the profile.

We are using EnsureDirState() to replace the source profiles. By always
generating the seccomp snippets, and including snap-seccomp in the profile
source, we ensure that the source file will be updated when there is a change in
libseccomp library version of the syscalls it supports.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>