tests: check that apt works before using it
The lxd test is using apt unconditionally even on core devices.

Nowdays the apt wrapper on core devices will error when run so
the lxd test needs an update to ensure it only uses apt if apt
is actually usable.

The ubuntu-core-apt test also needs an update to check that
the exit code is correct.