cmd/snap-confine: remove apparmor from sc_invocation
After a good night's sleep I realised I should not have put it there
in the first place. This will allow the invocatoin to collect and contain
only the parts of the arguments that are related to the invocation of
snap-confine and leave out certain things that have more complex state
(uids) or are an implementation detail (apparmor struct) out of it.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>