interfaces/apparmor: use available apparmor features
Historically snapd would fallback to "forced devmode", which is like
applying a very open and permissive confinement to all the systems that
do not implement full confinement. This is due to inability, at the
time, of ancient apparmor userspace, to even parse and compile the
regular snapd security template.

Over time we started allowing use of true partial confinement, that is
without the degradation to classic template, on openSUSE Tumbleweed.
That distribution was selected due to its rolling nature and the fact
that up-to-date apparmor userspace and kernel was available. The
experiment was further extended to cover Arch, along with the hardened
kernel that supports apparmor. This experiment was very successful,
showing that there are practical benefits in more complete sandbox
without any observable downsides.

Recently we reached a point where using the forced devmode fallback has
started to hurt the ecosystem, as people routinely develop snap packages
on Debian, openSUSE Leap or even on Ubuntu using mainline kernel. In
that situation they may not fully comprehend the snap interface system
and fail to use interfaces that are required under strict confinement.

While the complete journey towards explaining the state of confinement
is yet to be designed, we think it is important to use as many apparmor
features as are available, even if not all of the features we want are
present.

This patch will effectively start to enforce path based file protections
on Debian and openSUSE as well as any distribution running with apparmor
enabled. In practice it will shrink the perception delta between the
upstream kernel and the Ubuntu kernel down to fine-grained DBus
mediation and perhaps some of the network rules.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>