interfaces/apparmor: partial confinement on openSUSE
Historically snapd would fallback to "forced devmode", which is like
applying a very open and permissive confinement to all the systems that
do not implement full confinement. This is due to inability, at the
time, of ancient apparmor userspace, to even parse and compile the
regular snapd security template.

Over time we started allowing use of true partial confinement, that is
without the degradation to classic template, on openSUSE Tumblweed. That
distribution was selected due to its rolling nature and the fact that
up-to-date apparmor userspace and kernel was available.

Since Tumbleweed is used to build the rest of the SUSE family and the
required changes have since migrated to LEAP releases we can now extend
the fallback to handle both openSUSE (aka LEAP) as well as openSUSE
Tumbleweed, given the kernel version is recent enough.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>