tests: run smoke tests on (almost) pristine systems
We run the smoke suite in the autopkgtest backend with the
full prepare/restore logic. This is problematic as this
modifies the system in many ways, most importantly it will
rebuild the snapd deb. This is harmful in autopkgtest
because the deb under testing is already installed and
when we replace it with change what is tested.

There are multiple ways to fix this issue, this PR takes
an extreme way and removes all prepare/restore logic in
the smoke suite (and hence the autopkgtest) runs.

The smoke suite is also moved first to ensure that the
system has not had "prepare-suite" yet so the regular spread
tests are somewhat close to what runs inside the autopkgtest
environment. However there is still the "prepare-project"
bits that are run on regular spread but will be skipped on
autopkgtest. For this reason this PR also needs a test to
ensure we excercise the actual autopkgtest codepath and are
not running into unexpected autopkgtest failures at deb upload
time (probably via a nested test).