tests: add regression test for systemctl race fix
The systemd in UC16 and UC18 has a race that can cause
systemctl to hang forever when operations like "start"
are intereleaved with "systemctl daemon-reload"s.

This got fixed upstream in the systemd PR#8803

    https://github.com/systemd/systemd/pull/8803

And we backported this fix via the SRU

    https://bugs.launchpad.net/bugs/1819728

This PR adds a regression test that ensures the fix
works correctly.