cmd/snap-update-ns: use common load for UserProfileUpdate
This removes second half of the duplicate logic, the half being
the system mount profile logic, allowing for more unification
between the two algorithms (system and user).

We non also need to initialize private fields of UserProfileUpdate
so there is a new function NewUserProfileUpdate that allows this.

There is no "current" part of the per-user mount profiles since those are not
persisted yet. Much of the stage for that, however, is now set.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>