cmd/snap-confine: use fixed private tmp directory
When snap-confine prepares a fresh per-snap mount namespace
it constructs a private /tmp directory that is not shared with other
snaps.

As a part of this work snap-confine would create a randomly named
directory in /tmp/snap.$SNAP_NAME_xxxxx with the x characters replaced
by random unique name.

Those directories would never be recycled and would be re-created when
the mount namespace was to be discarded. This was accumulating data in
/tmp that would not be normally removed unless the machine had rebooted.
In addition whenever multiple temporary directories existed for a single
snap the user would be unable to easily determine which one is the right
one, presenting usability issues for applications that use data in /tmp
for whatever reason (e.g. temporary log files).

This patch changes the path to be fixed, always
/tmp/snap.$SNAP_INSTANCE_NAME/tmp. This opens the avenue towards snapd
managing that directory (e.g. to remove it on snap removal).

There is only one user-visible consequence: discarding the mount
namespace and re-running a snap application no longer gives the
application fresh /tmp directory. It is unlikely anything apart from
snapd test suite will observe this change though.

The test that was checking if the base snap for a given snap has changed
has now been updated not to rely on the re-creation of /tmp.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>
Related-To: https://bugzilla.suse.com/show_bug.cgi?id=1127368