overlord/snapstate: track time of postponed refreshes
When a refresh is attempted but for skipped because a particular snap is
busy then a time-stamp is recorded under snap state (refresh postponed
time). This time-stamp can be used to determine when a refresh is
mandatory, even if a snap is busy. This can prevent snaps from running
old revisions indefinitely simply because an app is open.

The patch adds the time-stamp to SnapState and ensures it is reset on
refresh. Somewhat arbitrarily the instant that this is done was picked
to be the "link-snap" task.

The timer is set and reset regardless of the refresh-app-awareness
feature flag because otherwise the state would go out-of-sync with
reality and this aspect of the change is not complex.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>