cmd/snap-confine: check for SNAP_INSTANCE_NAME early
The error message for unset SNAP_INSTANCE_NAME environment variable is
the only user interaction printed when snap-confine is invoked as a
program without any arguments. As such it is important that the error
message at least hints towards the solution of the problem.

At the same time the internals of sc_init_invocation() don't require the
value to come from the environment so it is sensible to use a dedicated
error message there.

This patch does both.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>