cmd/snap-update-ns: remove MountProfileUpdate.PerformChange
The MountProfileUpdate exists for decoupling the complexity behind each
method from the algorithm that operates on the interface implementations
which in turn simplifies testing.

This however muddies the water of what is actually being done,
especially when the Change.Perform method already has means of mocking
it with MockChangePerform.

As such, remove the method from the interface and use the Change.Perform
method directly.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>