cmd/snap-confine: fold sc_invocation_{fini,cleanup}
The _fini function is a destructor that does not concern itself with
heap or stack or bss allocation of the principal object. The cleanup
function is a combination of object destruction and memory deallocation
applicable for a GCC cleanup attribute extension which necessarily
marries the two concepts.

It was raised that in this case, of a stack allocated object, there is
no difference between the destructor and the cleanup function. As such I
will fold them into one function.

Separately, I will do a pass over snap-confine's memory handling to unify
it and give it a common naming scheme and propose the changes that will
be consistent across the tree.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>