cmd: prevent umask from breaking snap-run chain
It was brought to our attention that unconventional umask has
detrimental effect on the ability to run snaps. The principal reason is
that we have carefully crafted the mode of newly created directories and
files, we have not at all considered that they run with an umask value
inherited from the calling process.

There are two changes made here. Bulk of the creation of files and
directories is performed by snap-confine. There a single ad-hoc way of
handling umask in one part of the code is upgraded to a runtime-wide way
that runs around main. There's a small set of directories created
earlier, in snap-run, those gain local umask set/restore operation.

This patch introduces the snap-confine preservation context. A container
for process state that is clobbered by snap-confine that is preserved on
startup and restored just before executing the target application.

For now the context contains just the umask but I would like to store
the current working directory there in a follow-up patch. With umask
handled for snap-confine globally we can now remove the specialized
handling from setup_private_mount.

A new test ensures that even umask 777 is not preventing snap-confine
from working and that the mask is successfully applied to the invoked
application process.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>