snap-gdb-shim: switch to the SUDO_UID when available
Right now `snap run --gdb` will run everything as root.
This is a problem because most developers expect to debug
their programs as user. However we need to run as root
because the setuid snap-confine will not run with root
permissions when run under gdb.

The fix is to drop root again after snap-confine was run
but before the actual snap application starts.