cmd/snap-confine: always use fchdir + inner_cwd_fd
When originally implementing the first patch in this series I was
considering using the file descriptor for the current working directory
in the original mount namespace so I wanted to differentiate the cases
where the inode it points to now and before is different from that when
they are the same.

Over time I realised that I cannot use fchdir with the original
descriptor and implemented separate cases for fchdir + chdir but in
doing so I didn't realise that fchdir and chdir are equivalent here
because the path descriptor used is from the per-snap mount namespace
and that it is opened just moments before calling fchdir.

As such the whole fragment can be reduced to unconditional fchdir with
the inode check serving just as UX differentiation.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>