tests: try to be a little bit invariant
This stores a brief listing of some directories before a test runs,
and checks just before running the next test: there should be no
changes to those directories.

This will fail the test right after the one that messed up. Fixing
that would necessitate reworking the prepare/restore logic which is
too much work right now; the list of tests is there, so you just need
to look at the previous one.