seccomp: allow use of setgroups(0, NULL) by default
setgroups() takes a size parameter as the first argument and an array of
gid_t as the second argument. On Linux it is legal to specify a size of
'0' with NULL array as a convenience for dropping all supplemental
groups and this is safe to allow. Because of limitations in the kernel,
we cannot allow non-zero sizes with populated arrays since we cannot
mediate the array values. Allowing setgroups(0, NULL) will allow some
snaps to work unmodified while allowing others to at least drop all
groups. This opens the door for snapcraft-preload to replace these
unmediatable setgroups() calls with setgroups(0, NULL).