cmd/snap-confine: invert fallback/orig base snap name
The snap-confine invocation structure used to hold the base snap name
and the fallback base snap name that is sometimes used instead. This
required a helper to compute the effective base snap name, which
preferred the fallback if it was set.

This patch simplifies this logic by inverting the meaning. The base snap
name is always the one we are actually going to use, replacing the
effective base snap name helper function. In addition we have the
original base snap name that we remember. There's one specific place
that uses the original base snap name, namely the mount namespace
construction mode check, which decides if we should pivot into the base
snap or continue with the initial mount namespace as the starting point.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>