snap: support system-usernames for 'snap_daemon' user
For this initial phase where we support the snap_daemon user and use an
'implied assumes' such that when the snap specifies:

  system-usernames:
    snap_daemon: shared

we perform a uid and gid lookup for the user (ie, 'snap_daemon') to
ensure we can generate the seccomp policy correctly. If either or both
can't be found, alert the user that the user and group should be
created (creation of this user/group will be in a future commit).

In addition to the above yaml, the parser supports:

 system-usernames:
   <username1>: shared|private|external
   <username>:
     scope: shared|private|external
     attr1: ...
     attr2: ...

While the parser supports the above, the installation check will give an
appropriate error message that 'private' and 'external' are not yet
supported.