cmd/snap-confine: add is_same_base
The helper function can inspect /etc/os-release as visible in the
current mount namespace as and visible in the desired base snap and see
if the ID= and VERSION_ID= (if present) fields have the same value.

This helper can be used to detect when the base snap is not only
refreshed but changed from one base (e.g. core 16) to another (e.g.
core 18).

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>