cmd/snap-confine: add is_same_base
The helper function can inspect /usr/lib/os-release as visible in the
current mount namespace as and visible in the desired base snap and see
if the ID= and VERSION_ID= (if present) fields have the same value.

This helper can be used to detect when the base snap is not only
refreshed but changed from one base (e.g. core 16) to another (e.g.
core 18).

Note that looking at /etc/os-release is deceptive and wrong because it
is only "right" by accident, sometimes. The /etc directory comes from
the host and only recent-ish systems use a symbolic link to
../usr/lib/os-release. On some systems, for example Ubuntu 14.04, the
file is not a symbolic link and thus has the same value inside the
execution environment and outside of it.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>