cmd/snap-confine: discard mount namespace when base snap changes
When a snap is refreshed and the new snap uses a different base snap
than before (e.g. going from core to core18) then snap-confine must
discard the mount namespace, even at the cost of fracturing the view of
the mount namespace that individual app processes have.

To be able to do this we must know the name of the base snap in a given
mount namespace. This can be achieved in one of two ways: either by
being implicit and inspecting the mount namespace, trying to derive the
name of the base snap this way, or by being explicit, and storing the
name somewhere for us to read later.

The implicit path is technically possible but it is tricky and was
NACKed in early discussions around this problem. The explicit path is
taken instead and is described blow.

From now on, whenever snap-confine constructs a mount namespace it also
writes to /run/snapd/ns/snap.$SNAP_INSTANCE_NAME.info. The format of the
file is a sequence of key=value pairs, encoded one per line. For the
purpose of the bug only one line is written for now, base-snap-name=...

When snap-confine is invoked and checks for staleness of a mount
namespace it read the file again to access the name of the base snap
used. Whenever both values are available but differ then a base snap
transition is occurring and the mount namespace must be discarded, not
reused.

The set of messages sent across the eventfd grows from the old YES / NO
to MUST (force discard), SHOULD (discard if unused) and NO (reuse).  The
snap-discard-ns program is augmented to remove the .info file, if
present. The set of apparmor permissions are expanded to allow access to
said file.

Fixes: https://bugs.launchpad.net/snapd/+bug/1819875
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>