cmd/snap-confine: discard mount namespace when base snap changes
When a snap is refreshed and the new snap uses a different base snap
than before (e.g. going from core to core18) then snap-confine must
discard the mount namespace, even at the cost of fracturing the view of
the mount namespace that individual app processes have.

Since snap-confine doesn't really know the name of the old base snap, it
can instead use a trick based on analysis of the os-release file. When
the os-release file's ID and VERSION_ID, as seen from the point of view
of the old (preserved) mount namespace differ from those of the desired
base snap then the snap mount namespace is forcibly discarded and
recreated.

The set of messages sent across the eventfd grows from the old YES / NO
to MUST (force discard), SHOULD (discard if unused) and NO (reuse).

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>