testutil: make mocked command work with shellcheck from snaps
When shellcheck is running from a snap it will get a private /tmp. Scripts are
generated in a workdir that is places under /tmp, but inside the host mount
namespace, thus the file will not be accessible by shellcheck.

Work around that by feeding the script to be checked through stdin.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>