cmd/snap-confine: goto void on E{PERM,ACCESS} orig dir
Maciej Borzęcki rightfully noted that we can fail to open the path
(using O_PATH flag) when we don't have permissions to navigate the
directory chain. Instead of dying like we do on most errors, we will now
go to the void directory. This is more in line with similar case when
fchdir fails to change the directory for permission reasons.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>