daemon: add RootOnly flag to commands
Without this change, endpoints that needed the user to be actual root
(as opposed to gaining admin via login or polkit) would check ucrednet
directly.

This change adds a RootOnly flag to commands, and uses it for those
commands that need it, moving the checks to canAccess together with
the rest of access checks.