cmd/snap-confine: reject crafted /tmp/snap.$SNAP_NAME
In case an attacker crafts a base directory name /tmp/snap.$SNAP_NAME
with inappropriate ownership or permissions snap-confine will now detect
such attempt and refuse to use it.

The test that probes /tmp was a bit quaint so it got a face-lift as well
as some new checks for the features added here.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>