cmd/libsnap: remove fringe error function
The codebase before snap-confine had a family of functions such as
debug, error and die. The error function simply printed a message to
stderr. We have exactly one use of that function:

        sc_error *err SC_CLEANUP(sc_cleanup_error) = NULL;
        snap_context =
            sc_cookie_get_from_snapd(invocation.snap_instance, &err);
        if (err != NULL) {
            error("%s\n", sc_error_msg(err));
        }

When we cannot get a cookie from snapd, we print the error and carry on.
This is not really useful, the possible errors, apart from memory
allocation and I/O, are simply the absence of the cookie.

This code was originally written this way because ancient snapd was not
writing the cookies and, when upgrading, the previously fatal error
would prevent apps from running.

The error message is confusing and doesn't help: users cannot do
anything about it. As such, don't print the error at all.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>