snap: keep side state for implicit hook binding
When binding implicit hooks to plugs and slots we are inevitably
modifying plugs and slots with new hook associations. We also use hook
associations to know which interfaces to skip so this is clearly broken.

Use local state to figure out which plugs or slots are already bound and
should not be added to implicit hooks.

Tests were adjusted to show this is working correctly. I tweaked the
test code layout a little to move the YAML constants next to individual
test cases. I also fixed a small bug that was using the same snap name
inside the single test case, creating spurious implicit hooks. It was
masked by the use of the same hook name (implicit) in both
sub-test-cases before.

Thanks to Samuele for spotting the problem.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>