tests/main/user-data-handling: do not glob for home dirs
Do not attempt to use globs when trying to find home directories, as this may
pick up directories that are not homedirs.

The spread data is dropped under /home/gopath, the test proceeds to create some
data inside, but since it is not a home directory, SELinux context transitions
do not take place (at least on CentOS) and mock data is left with incorrect
labeling.

Instead of guessing the home directories, expect a test and root users.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>