snap: track scoped plugs slots, discard toplevel concept
Third iteration of the attempt to fix the problem when implicit hooks
would get incorrect plug and slot associations. In the past we used to
track top-level plugs and slots. Those are naturally plugs and slots
defined at the top level of the YAML structure. Such plugs and slots,
unless explicitly associated (or scoped) to a specific application or
hook are implicitly associated with all of them.

Due to the fact that there was no extra state, scoped plugs and slots
were determined by checking if they have any apps or hooks defined on
them at the particular moment when implicit hook handling was done. If
it was, however, associated with an app it would no longer be a
candidate for implicitly-defined hooks.

To solve this situation, hopefully with less tricky code, the following
is done. PlugInfo and SlotInfo gain a new public attribute, Scoped, that
is set whenever an app or hook refers to a plug or slot. When automatic
plug-or-slot to app-or-hook associations are performed they only look at
the Scoped flag to know what to do.

A pair of extra tests covering this situation is also added.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>