osutil: make CommandFromCore honor the snapd snap
The CommandFromCore helper is currently always looking at the
core snap. With the snapd snap we need to refactor it so that
it uses the snapd snap if available and just falls back to the
core snap.

This PR is a RFC with the minimal amount of code to do this and
some new integration tests. It also adds the xdelta3 and mksquashfs
tools to the snapd snap.