cmdutil: move osutil.CommandFromCore to cmdutil.CommandFromSystemSnap
The osutil.CommandFromCore is not in the rightplace and it also does
not take the snapd snap into account. This commit moves it to the
cmd/cmdutil package (cannot be in cmd because of an import cycle
with cmd/snapinfo.go).

It also tweaks the arguments slightly (no need to pass the
snapMountDir anymore).