adjust checkSystemUsers() to lookup the system-user with helpful error
For this initial phase where we support the LSB-required daemon user,
use an 'implied assumes' such that when the snap specifies:

 system-users: [ daemon ]

we perform a uid and gid lookup for the user (ie, 'daemon') to ensure we
can generate the seccomp policy correctly. If either or both can't be
found, give an appropriate useradd invocation to create a system user
with a pointer to the useradd man page.