Change the snap used for snapshot-basic test
This test is using the test-snapd-rsync snap as part of the test.
It causes an issue on core18 because as it is included in the
SKIP_REMOVE_SNAPS list when the next test calls the reset.sh the snap is
not removed but then fails to remove the core snap (now reset.sh fails
when it can't remove the base/core snap).