cmd/snap-confine: move privs handling code one file
Internally snap-confine has a number of places where is temporarily
drops permissions, raises permissions and finally drops them
permanently. As I'm about to refactor that so that it participates with
the preserved-process-state system but since it's a bit complex I
thought it would help to bring it under one place for easier analysis of
the upcoming changes.

This patch adds snap-confine-privs.c with a functions containing
fragments of each of the user and group ID handling parts of snap
confine. The new functions are named after the place they were extracted
from and documented to the best of my understanding. I'm honestly doing
this to wrap *my* head around it completely and have fewer surprises
later.

The new file is temporary, once the preserved process state refactoring
is merged it can be removed as it should become empty.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>