don't allow 'system-users' when snap-seccomp is built with libseccomp < 2.4
libseccomp < 2.4 generates unpredicatable BPFs where at runtime the
policy enforcement for argument filtered rules is sometimes wrong. The
BPF generation is known to be sensitive to rule order and while the PFC
is correct, a disassembly of the BPF indicates problematic logic that is
also sensitive to the runtime environment. Incorrect enforcement was
observed with secondary architectures and differing kernel versions
across different distros. Please see PR 6681 for details.

libseccomp 2.4 includes a rewrite of the BPF to, in part, better handle
argument filtering and it now generates predictable BPFs (as seen from
the PFCs and BPF disassemblies) without the problematic logic that is
sensitive to runtime environments. (As an aside, it also orders the
syscalls based on priorities, which should yield performance gains for
snaps with heavy syscall use). As such, require that the 'system-users'
features be dependent on snap-seccomp being compiled against libseccomp
>= 2.4.

References:
https://github.com/snapcore/snapd/pull/6681