devicestate: make Remodel() return a state.Change
The devicestate.Remodel() call returns a []*state.TaskSet right
now. This is the usual pattern we use but for Remodel() its
not the right one because Remodel() cannot be composed with
other changes (doing that would be dangerous because of the
e.g. the task serialization inside remodel). Having a change
also means we can add the new model directly to the change
which will be useful when remodeling to different kernels
or (boot) bases as the validation code will need access to
the new model.

So this PR implements the switch from []*state.TaskSet to
*change. Followups will use it to add the model to the change
and to remodel to a different kernel/base.