snapstate: auto-install snapd when needed
When installing a snap that does not require "core" on a fresh
system, currently we do nothing special. This means however that
on such systems there is no "core" or "snapd" snap installed.

This means that no system interfaces are available. This PR
fixes this by auto-installing the snapd snap if its missing.

As a side effect, "snap install snapd" on classic is no longer
blocked (it still is on core).